+++ backends-clickhouse/src/test/scala/org/apache/spark/sql/extension/GlutenCustomerExtensionSuite.scala
@@ -17,11 +17,16 @@
 package org.apache.spark.sql.extension
 
 import org.apache.gluten.config.GlutenConfig
+import org.apache.gluten.sql.shims.SparkShimLoader
 
 import org.apache.spark.SparkConf
 import org.apache.spark.sql.GlutenSQLTestsTrait
+import org.apache.spark.sql.SparkSession
+import org.apache.spark.sql.catalyst.rules.Rule
+import org.apache.spark.sql.execution.{FileSourceScanExec, SparkPlan}
+import org.apache.spark.sql.test.SharedSparkSession
 
-class GlutenCustomerExtensionSuite extends GlutenSQLTestsTrait {
+class GlutenCustomerExtensionSuite extends SharedSparkSession {
   // These configs only take effect on ClickHouse backend.
   private val ExtendedColumnarTransformRulesKey =
     "spark.gluten.sql.columnar.extended.columnar.transform.rules"
@@ -49,8 +54,32 @@ class GlutenCustomerExtensionSuite extends GlutenSQLTestsTrait {
       val testFileSourceScanExecTransformer = df.queryExecution.executedPlan.collect {
         case f: TestFileSourceScanExecTransformer => f
       }
-      assert(!testFileSourceScanExecTransformer.isEmpty)
-      assert(testFileSourceScanExecTransformer(0).nodeNamePrefix.equals("TestFile"))
+      assert(testFileSourceScanExecTransformer.nonEmpty)
+      assert(testFileSourceScanExecTransformer.head.nodeNamePrefix.equals("TestFile"))
     }
+  }
+}
+
+case class CustomerColumnarPreRules(session: SparkSession) extends Rule[SparkPlan] {
+
+  override def apply(plan: SparkPlan): SparkPlan = plan.transformDown {
+    case fileSourceScan: FileSourceScanExec =>
+      val transformer = new TestFileSourceScanExecTransformer(
+        fileSourceScan.relation,
+        SparkShimLoader.getSparkShims.getFileSourceScanStream(fileSourceScan),
+        fileSourceScan.output,
+        fileSourceScan.requiredSchema,
+        fileSourceScan.partitionFilters,
+        fileSourceScan.optionalBucketSet,
+        fileSourceScan.optionalNumCoalescedBuckets,
+        fileSourceScan.dataFilters,
+        fileSourceScan.tableIdentifier,
+        fileSourceScan.disableBucketedScan
+      )
+      if (transformer.doValidate().ok()) {
+        transformer
+      } else {
+        plan
+      }
   }
 }
